gis/requirements.txt
dask
geopandas
odc-stac
pyproj
pystac
pystac-client
rasterio
rioxarray
shapely
xarray
xarray-spatial
